requirements.txt
absl-py==1.0.0
astroid==2.4.2
astunparse==1.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.11.0
dnspython==2.0.0
Flask==1.1.2
Flask-SSLify==0.1.5
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
gunicorn==20.0.4
h5py==3.5.0
html5lib==1.1
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
libclang==12.0.0
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.3
mccabe==0.6.1
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pymongo==3.11.2
pyparsing==3.0.5
PySocks==1.7.1
pytesseract==0.3.6
python-dateutil==2.8.2
python-dotenv==0.15.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7.2
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
stem==1.8.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
